requirements.txt
attrs==20.3.0
Django==2.2.5
djangorestframework==3.9.2
gunicorn==20.0.4
importlib-metadata==3.7.3
iniconfig==1.1.1
packaging==20.9
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
zipp==3.4.1
